Big refactor to remove API-based models from TripPlan flow

+++ DigiTransit10/Models/Favorite.cs
@@ -55,7 +55,7 @@ namespace DigiTransit10.Models
     public struct FavoriteRoutePlace
     {
         public string Name { get; set; }
-        public float Lat { get; set; }
-        public float Lon { get; set; }
+        public double Lat { get; set; }
+        public double Lon { get; set; }
     }
 }
